pyproject.toml
[project]
name = "biblio"
version = "0.1.0"
description = ""
authors = [
    {name = "msarotti",email = "marco.sarotti@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0.0"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-browser-reload (>=1.18.0,<2.0.0)",
    "django-tailwind (>=3.8.0,<4.0.0)"
]

[tool.poetry]
packages = [{include = "biblio", from = "src"}]


[tool.poetry.group.dev.dependencies]
debugpy = "^1.8.13"
flake8 = "^7.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
